move app-specific fixtures inside app
conftest defines test fixtures e.g. `assessment` - some of these are
specific to an app-version, so they should be in the app directory
